Move API call to client class
